[AI Generated] Integrated new functionalities into existing knowledge_ingestion.py and removed redundant files
