--- agent/ethical_checks.py
@@ -1,14 +1,0 @@
-class EthicalChecks:
-    @staticmethod
-    def complies_with_guidelines(content):
-        # Placeholder for ethical guideline checks
-        return not any(word in content.lower() for word in ['hate', 'violence', 'discrimination'])
-
-    @staticmethod
-    def is_safe(content):
-        # Placeholder for safety checks
-        return 'unsafe' not in content.lower()
-
-    @staticmethod
-    def validate(content):
-        return EthicalChecks.complies_with_guidelines(content) and EthicalChecks.is_safe(content)
